+++ src/conv.rs
@@ -3,7 +3,8 @@ use wgc::Label;
 use crate::native;
 use crate::utils::{make_slice, ptr_into_label, ptr_into_pathbuf};
 use crate::{follow_chain, map_enum};
-use std::{borrow::Cow, ffi::CStr, slice};
+use std::num::{NonZeroU32, NonZeroU64};
+use std::{borrow::Cow, ffi::CStr};
 
 // Ultimately map_load_op and map_store_op should be able to use the map_enum!
 // macro, but not until wgc supports "Undefined" as a valid value. For now the
@@ -1132,55 +1133,184 @@ pub fn to_native_composite_alpha_mode(
 pub fn map_bind_group_entry<'a>(
     entry: &'a native::WGPUBindGroupEntry,
     extras: Option<&native::WGPUBindGroupEntryExtras>,
-) -> Option<wgc::binding_model::BindGroupEntry<'a>> {
-    let entry_extras = extras?;
-    unsafe {
-        if entry_extras.textureViews.as_ref().is_some() {
-            let texture_views = slice::from_raw_parts(
-                entry_extras.textureViews,
-                usize::try_from(entry_extras.textureViewCount).unwrap(),
-            )
-            .iter()
-            .map(|view_ptr| view_ptr.as_ref().unwrap().id)
-            .collect(); //Vec::new();
-
-            Some(wgc::binding_model::BindGroupEntry {
+) -> wgc::binding_model::BindGroupEntry<'a> {
+    if let Some(buffer) = unsafe { entry.buffer.as_ref() } {
+        return wgc::binding_model::BindGroupEntry {
+            binding: entry.binding,
+            resource: wgc::binding_model::BindingResource::Buffer(
+                wgc::binding_model::BufferBinding {
+                    buffer_id: buffer.id,
+                    offset: entry.offset,
+                    size: match entry.size {
+                        0 => panic!("invalid size"),
+                        WGPU_WHOLE_SIZE => None,
+                        _ => Some(unsafe { NonZeroU64::new_unchecked(entry.size) }),
+                    },
+                },
+            ),
+        };
+    } else if let Some(sampler) = unsafe { entry.sampler.as_ref() } {
+        return wgc::binding_model::BindGroupEntry {
+            binding: entry.binding,
+            resource: wgc::binding_model::BindingResource::Sampler(sampler.id),
+        };
+    } else if let Some(texture_view) = unsafe { entry.textureView.as_ref() } {
+        return wgc::binding_model::BindGroupEntry {
+            binding: entry.binding,
+            resource: wgc::binding_model::BindingResource::TextureView(texture_view.id),
+        };
+    } else if let Some(extras) = extras {
+        if let Some(texture_views) = unsafe { extras.textureViews.as_ref() } {
+            let arr = make_slice(texture_views, extras.textureViewCount)
+                .iter()
+                .map(|v| {
+                    unsafe { v.as_ref() }
+                        .expect("invalid texture views for bind group entry extras")
+                        .id
+                })
+                .collect();
+            return wgc::binding_model::BindGroupEntry {
                 binding: entry.binding,
-                resource: wgc::binding_model::BindingResource::TextureViewArray(texture_views),
-            })
-        } else if entry_extras.samplers.as_ref().is_some() {
-            let samplers = slice::from_raw_parts(
-                entry_extras.samplers,
-                usize::try_from(entry_extras.samplerCount).unwrap(),
-            )
-            .iter()
-            .map(|sampler_ptr| sampler_ptr.as_ref().unwrap().id)
-            .collect(); //Vec::new();
-
-            Some(wgc::binding_model::BindGroupEntry {
+                resource: wgc::binding_model::BindingResource::TextureViewArray(arr),
+            };
+        } else if let Some(samplers) = unsafe { extras.samplers.as_ref() } {
+            let arr = make_slice(samplers, extras.samplerCount)
+                .iter()
+                .map(|v| {
+                    unsafe { v.as_ref() }
+                        .expect("invalid sampler for bind group entry extras")
+                        .id
+                })
+                .collect();
+            return wgc::binding_model::BindGroupEntry {
                 binding: entry.binding,
-                resource: wgc::binding_model::BindingResource::SamplerArray(samplers),
-            })
-        } else if entry_extras.buffers.as_ref().is_some() {
-            let buffers = slice::from_raw_parts(
-                entry_extras.buffers,
-                usize::try_from(entry_extras.bufferCount).unwrap(),
-            )
-            .iter()
-            .map(|buffer_ptr| wgc::binding_model::BufferBinding {
-                buffer_id: buffer_ptr.as_ref().unwrap().id,
-                offset: entry.offset,
-                size: std::num::NonZeroU64::new(entry.size),
-            })
-            .collect();
-
-            Some(wgc::binding_model::BindGroupEntry {
+                resource: wgc::binding_model::BindingResource::SamplerArray(arr),
+            };
+        } else if let Some(buffers) = unsafe { extras.buffers.as_ref() } {
+            let arr = make_slice(buffers, extras.bufferCount)
+                .iter()
+                .map(|v| wgc::binding_model::BufferBinding {
+                    buffer_id: unsafe { v.as_ref() }
+                        .expect("invalid buffers for bind group entry extras")
+                        .id,
+                    offset: entry.offset,
+                    size: std::num::NonZeroU64::new(entry.size),
+                })
+                .collect();
+            return wgc::binding_model::BindGroupEntry {
                 binding: entry.binding,
-                resource: wgc::binding_model::BindingResource::BufferArray(buffers),
-            })
-        } else {
-            None
+                resource: wgc::binding_model::BindingResource::BufferArray(arr),
+            };
         }
+    }
+
+    panic!("invalid bind group entry for bind group descriptor");
+}
+
+#[inline]
+pub fn map_bind_group_layout_entry<'a>(
+    entry: &'a native::WGPUBindGroupLayoutEntry,
+    extras: Option<&native::WGPUBindGroupLayoutEntryExtras>,
+) -> wgt::BindGroupLayoutEntry {
+    let is_buffer = entry.buffer.type_ != native::WGPUBufferBindingType_Undefined;
+    let is_sampler = entry.sampler.type_ != native::WGPUSamplerBindingType_Undefined;
+    let is_texture = entry.texture.sampleType != native::WGPUTextureSampleType_Undefined;
+
+    let is_storage_texture =
+        entry.storageTexture.access != native::WGPUStorageTextureAccess_Undefined;
+
+    let ty = if is_texture {
+        wgt::BindingType::Texture {
+            sample_type: match entry.texture.sampleType {
+                native::WGPUTextureSampleType_Float => {
+                    wgt::TextureSampleType::Float { filterable: true }
+                }
+                native::WGPUTextureSampleType_UnfilterableFloat => {
+                    wgt::TextureSampleType::Float { filterable: false }
+                }
+                native::WGPUTextureSampleType_Depth => wgt::TextureSampleType::Depth,
+                native::WGPUTextureSampleType_Sint => wgt::TextureSampleType::Sint,
+                native::WGPUTextureSampleType_Uint => wgt::TextureSampleType::Uint,
+                _ => panic!("invalid sample type for texture binding layout"),
+            },
+            view_dimension: match entry.texture.viewDimension {
+                native::WGPUTextureViewDimension_1D => wgt::TextureViewDimension::D1,
+                native::WGPUTextureViewDimension_2D => wgt::TextureViewDimension::D2,
+                native::WGPUTextureViewDimension_2DArray => wgt::TextureViewDimension::D2Array,
+                native::WGPUTextureViewDimension_Cube => wgt::TextureViewDimension::Cube,
+                native::WGPUTextureViewDimension_CubeArray => wgt::TextureViewDimension::CubeArray,
+                native::WGPUTextureViewDimension_3D => wgt::TextureViewDimension::D3,
+                _ => panic!("invalid texture view dimension for texture binding layout"),
+            },
+            multisampled: entry.texture.multisampled != 0,
+        }
+    } else if is_sampler {
+        match entry.sampler.type_ {
+            native::WGPUSamplerBindingType_Filtering => {
+                wgt::BindingType::Sampler(wgt::SamplerBindingType::Filtering)
+            }
+            native::WGPUSamplerBindingType_NonFiltering => {
+                wgt::BindingType::Sampler(wgt::SamplerBindingType::NonFiltering)
+            }
+            native::WGPUSamplerBindingType_Comparison => {
+                wgt::BindingType::Sampler(wgt::SamplerBindingType::Comparison)
+            }
+            _ => panic!("invalid sampler binding type for sampler binding layout"),
+        }
+    } else if is_storage_texture {
+        wgt::BindingType::StorageTexture {
+            access: match entry.storageTexture.access {
+                native::WGPUStorageTextureAccess_WriteOnly => wgt::StorageTextureAccess::WriteOnly,
+                _ => {
+                    panic!("invalid storage texture access for storage texture binding layout")
+                }
+            },
+            format: map_texture_format(entry.storageTexture.format)
+                .expect("invalid texture format for storage texture binding layout"),
+            view_dimension: match entry.storageTexture.viewDimension {
+                native::WGPUTextureViewDimension_1D => wgt::TextureViewDimension::D1,
+                native::WGPUTextureViewDimension_2D => wgt::TextureViewDimension::D2,
+                native::WGPUTextureViewDimension_2DArray => wgt::TextureViewDimension::D2Array,
+                native::WGPUTextureViewDimension_Cube => wgt::TextureViewDimension::Cube,
+                native::WGPUTextureViewDimension_CubeArray => wgt::TextureViewDimension::CubeArray,
+                native::WGPUTextureViewDimension_3D => wgt::TextureViewDimension::D3,
+                _ => {
+                    panic!("invalid texture view dimension for storage texture binding layout")
+                }
+            },
+        }
+    } else if is_buffer {
+        wgt::BindingType::Buffer {
+            ty: match entry.buffer.type_ {
+                native::WGPUBufferBindingType_Uniform => wgt::BufferBindingType::Uniform,
+                native::WGPUBufferBindingType_Storage => {
+                    wgt::BufferBindingType::Storage { read_only: false }
+                }
+                native::WGPUBufferBindingType_ReadOnlyStorage => {
+                    wgt::BufferBindingType::Storage { read_only: true }
+                }
+                _ => panic!("invalid buffer binding type for buffer binding layout"),
+            },
+            has_dynamic_offset: entry.buffer.hasDynamicOffset != 0,
+            min_binding_size: {
+                assert_ne!(
+                    entry.buffer.minBindingSize, WGPU_WHOLE_SIZE,
+                    "invalid min binding size for buffer binding layout, use 0 instead"
+                );
+
+                NonZeroU64::new(entry.buffer.minBindingSize)
+            },
+        }
+    } else {
+        panic!("invalid bind group layout entry for bind group layout descriptor");
+    };
+
+    wgt::BindGroupLayoutEntry {
+        ty,
+        binding: entry.binding,
+        visibility: wgt::ShaderStages::from_bits(entry.visibility)
+            .expect("invalid visibility for bind group layout entry"),
+        count: extras.map(|v| NonZeroU32::new(v.count)).flatten(),
     }
 }
